+++ test/test_cancel_queries.py
@@ -1,3 +1,5 @@
+from test.util import VerifyingQueryRunner
+
 from aio_overpass import Client
 
 import pytest
@@ -11,7 +13,7 @@ URL_KILL = "https://overpass-api.de/api/kill_my_queries"
 async def test_cancel_no_queries():
     body = ""
 
-    c = Client()
+    c = Client(runner=VerifyingQueryRunner())
 
     with aioresponses() as m:
         m.get(
@@ -34,7 +36,7 @@ Killing query (pid 7118) from IP 2a02:8108:41c0:2d95:109a:1b1:dacd:c917 ...
 Done!
     """
 
-    c = Client()
+    c = Client(runner=VerifyingQueryRunner())
 
     with aioresponses() as m:
         m.get(
@@ -59,7 +61,7 @@ Killing query (pid 7118) from IP 2a02:8108:41c0:2d95:109a:1b1:dacd:c917 ...
 Done!
     """
 
-    c = Client()
+    c = Client(runner=VerifyingQueryRunner())
 
     with aioresponses() as m:
         m.get(
@@ -84,7 +86,7 @@ Killing query (pid 7119) from IP 2a02:8108:41c0:2d95:109a:1b1:dacd:c917 ...
 Done!
     """
 
-    c = Client()
+    c = Client(runner=VerifyingQueryRunner())
 
     with aioresponses() as m:
         m.get(